requirements.txt
streamlit
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.3.2
scikit-surprise==1.1.3
plotly==5.18.0
requests==2.31.0
joblib==1.3.2
